+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - compatibility with `python3.6`
+- no longer report button automator's battery percentage on deprecated topic
+  `homeassistant/cover/switchbot/+/battery-percentage`
+  (use `homeassistant/switch/switchbot/+/battery-percentage` instead, see `v2.1.0`)
 
 ## [2.2.0] - 2021-10-23
 ### Added